[Core] add types.h and replace old uint32_t into u32 etc

+++ source/shared/ccore/buffer.h
@@ -7,9 +7,7 @@ extern "C"
 {
 #endif // __cplusplus
 	
-#include <inttypes.h>
-#include <stdbool.h>
-#include <stddef.h>
+#include "types.h"
 #include "api.h"
 	
 typedef struct fc_alloc_callbacks_t fc_alloc_callbacks_t;
@@ -17,7 +15,7 @@ typedef struct fc_alloc_callbacks_t fc_alloc_callbacks_t;
 typedef struct fc_binary_buffer_t
 {
 	void* pData;
-	size_t size;
+	u64 size;
 } fc_binary_buffer_t;
 
 CCORE_API bool fc_load_binary_file_into_binary_buffer(const char* path, fc_binary_buffer_t* pBuffer, fc_alloc_callbacks_t* pAllocCallbacks);
@@ -31,13 +29,13 @@ typedef struct fc_binary_buffer_stream_t
 } fc_binary_buffer_stream_t;
 
 CCORE_API void fc_init_binary_buffer_stream(const fc_binary_buffer_t* buffer, fc_binary_buffer_stream_t* outStream);
-CCORE_API uint32_t fc_read_binary_buffer(fc_binary_buffer_stream_t* stream, uint32_t numBytes, void* output);
-CCORE_API uint32_t fc_peek_binary_buffer(fc_binary_buffer_stream_t* stream, uint32_t numBytes, void* output);
+CCORE_API u32 fc_read_binary_buffer(fc_binary_buffer_stream_t* stream, u32 numBytes, void* output);
+CCORE_API u32 fc_peek_binary_buffer(fc_binary_buffer_stream_t* stream, u32 numBytes, void* output);
 
 typedef struct fc_text_buffer_t
 {
 	char* pData;
-	size_t size;
+	u64 size;
 } fc_text_buffer_t;
 	
 bool fc_load_text_file_into_text_buffer(const char* path, fc_text_buffer_t* pBuffer, fc_alloc_callbacks_t* pAllocCallbacks);
